backend/foodgram/requirements.txt
django==3.2.18
djoser==2.1.0
djangorestframework==3.14.0
djangorestframework_simplejwt==4.8.0
Pillow==9.4.0
django-import-export==3.1.0
django-filter==22.1
